pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kotoba"
version = "0.0.1"
description = "Web testing tool using Japanese natural language"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "kotoba Development Team"},
]
maintainers = [
    {name = "0xkaz"},
]
keywords = ["japanese", "nlp", "web", "testing", "playwright", "llm", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Natural Language :: Japanese",
    "Natural Language :: English",
]

dependencies = [
    "playwright>=1.38.0",
    "transformers>=4.34.0",
    "torch>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
gpu = [
    "torch>=2.0.0",
    "accelerate>=0.20.0",
]

[project.urls]
Homepage = "https://github.com/0xkaz/kotoba"
Repository = "https://github.com/0xkaz/kotoba"
Issues = "https://github.com/0xkaz/kotoba/issues"
Changelog = "https://github.com/0xkaz/kotoba/blob/main/CHANGELOG.md"

[project.scripts]
kotoba = "kotoba.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["kotoba*"]

[tool.setuptools.package-data]
kotoba = ["configs/*.yaml"]

[tool.ruff]
line-length = 120
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]